backend/flask/requirements.txt
APScheduler==3.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.7.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
requirements-parser==0.2.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
uWSGI==2.0.18
Werkzeug==0.15.2
